Refactor code by moving View-related logic to the View class
